requirements.txt
fastapi
uvicorn[standard]
websockets
mysql-connector-python
python-dotenv
jinja2